+++ extensions/drupal-org/CHANGELOG.md
@@ -1,5 +1,9 @@
 # drupal.org Changelog
 
+## [✨ New Feature] - 2024-05-08
+
+- Added command to search Drupal's API (api.drupal.org)
+
 ## [Initial Version] - 2022-12-28
 
 - Added command to search for Drupal records